+++ lib/createGltf.js
