+++ lib/loadImage.js
