+++ lib/loadMtl.js
