+++ lib/loadObj.js
@@ -4,8 +4,8 @@ var path = require('path');
 var Promise = require('bluebird');
 
 var ArrayStorage = require('./ArrayStorage');
-var loadImage = require('./image');
-var loadMtl = require('./mtl');
+var loadImage = require('./loadImage');
+var loadMtl = require('./loadMtl');
 var readLines = require('./readLines');
 
 var ComponentDatatype = Cesium.ComponentDatatype;
