+++ lib/obj2gltf.js
@@ -4,8 +4,8 @@ var fsExtra = require('fs-extra');
 var GltfPipeline = require('gltf-pipeline').Pipeline;
 var path = require('path');
 var Promise = require('bluebird');
-var createGltf = require('./gltf');
-var loadObj = require('./obj');
+var createGltf = require('./createGltf');
+var loadObj = require('./loadObj');
 var writeUris = require('./writeUris');
 
 var fsExtraOutputJson = Promise.promisify(fsExtra.outputJson);
@@ -14,7 +14,7 @@ var defaultValue = Cesium.defaultValue;
 var defined = Cesium.defined;
 var DeveloperError = Cesium.DeveloperError;
 
-module.exports = convert;
+module.exports = obj2gltf;
 
 /**
  * Converts an obj file to a glTF file.
@@ -37,8 +37,8 @@ module.exports = convert;
  * @param {Boolean} [options.secure=false] Prevent the converter from reading image or mtl files outside of the input obj directory.
  * @param {Logger} [options.logger] A callback function for handling logged messages. Defaults to console.log.
  */
-function convert(objPath, gltfPath, options) {
-    var defaults = convert.defaults;
+function obj2gltf(objPath, gltfPath, options) {
+    var defaults = obj2gltf.defaults;
 
     options = defaultValue(options, {});
     var binary = defaultValue(options.binary, defaults.binary);
@@ -113,7 +113,7 @@ function convert(objPath, gltfPath, options) {
         })
         .then(function(gltf) {
             if (bypassPipeline) {
-                return convert._outputJson(gltfPath, gltf);
+                return obj2gltf._outputJson(gltfPath, gltf);
             } else {
                 return GltfPipeline.processJSONToDisk(gltf, gltfPath, pipelineOptions);
             }
@@ -121,9 +121,9 @@ function convert(objPath, gltfPath, options) {
 }
 
 /**
- * Default values that will be used when calling convert(options) unless specified in the options object.
+ * Default values that will be used when calling obj2gltf(options) unless specified in the options object.
  */
-convert.defaults = {
+obj2gltf.defaults = {
     /**
      * Gets or sets whether the model will be saved as binary glTF.
      * @type Boolean
@@ -210,7 +210,7 @@ convert.defaults = {
  *
  * @private
  */
-convert._outputJson = fsExtraOutputJson;
+obj2gltf._outputJson = fsExtraOutputJson;
 
 /**
  * A callback function that logs messages.
