+++ package.json
@@ -42,6 +42,7 @@
     "istanbul": "^0.4.5",
     "jasmine": "^2.5.3",
     "jasmine-spec-reporter": "^3.2.0",
+    "jsdoc": "^3.4.3",
     "jshint": "^2.9.4",
     "jshint-stylish": "^2.2.1",
     "open": "^0.0.5",
